requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.11.0
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
bm3d==4.0.3
bm4d==4.2.5
boto3==1.38.15
botocore==1.38.15
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.0
clint==0.5.1
clip-score==0.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
diffusers==0.35.2
dill==0.3.8
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
grpcio==1.73.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
INNLab==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
loguru==0.7.3
lpips==0.1.4
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prettytable==3.17.0
propcache==0.3.1
protobuf==6.31.1
psutil==7.1.3
pyarrow==20.0.0
pycryptodome==3.23.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.12.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.35.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.21.1
torch @ file:///home/hdda/zyz/pytorch/torch-2.4.1%2Bcu124-cp310-cp310-linux_x86_64.whl#sha256=464cb998fa46d34317b10811dad88486aa8aeb96ebbcefc3ed3f00ddf6c3249b
torchmetrics==0.11.0
torchvision @ file:///home/hdda/zyz/pytorch/torchvision-0.19.1%2Bcu124-cp310-cp310-linux_x86_64.whl#sha256=a96c5533c6914afe1ded9e92ad8981e2143f5ed89cd811b842eda1aff01b261c
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.0.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
